requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
psycopg==3.3.3
psycopg-binary==3.3.3
sqlparse==0.5.5
tzdata==2025.3